Avoid fetching identity when using noop analytics reporter
