+++ internal/analytics/default_reporter.go
@@ -1,10 +1,23 @@
 package analytics
 
 import (
+	"context"
+	"encoding/json"
+	"errors"
 	"fmt"
+	"github.com/kubeshop/botkube/pkg/version"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/client-go/kubernetes"
+	"os"
+	"path/filepath"
 
 	segment "github.com/segmentio/analytics-go"
 	"github.com/sirupsen/logrus"
+)
+
+const (
+	kubeSystemNSName  = "kube-system"
+	analyticsFileName = "analytics.yaml"
 )
 
 var (
@@ -45,7 +58,21 @@ func NewDefaultReporter(log logrus.FieldLogger) (*DefaultReporter, CleanupFn, er
 		nil
 }
 
-func (r *DefaultReporter) RegisterIdentity(identity Identity) error {
+func (r *DefaultReporter) RegisterCurrentIdentity(ctx context.Context, k8sCli kubernetes.Interface, cfgDir string) error {
+	currentIdentity, err := r.load(ctx, k8sCli, cfgDir)
+	if err != nil {
+		return fmt.Errorf("while loading current identity: %w", err)
+	}
+
+	err = r.registerIdentity(currentIdentity)
+	if err != nil {
+		return fmt.Errorf("while registering identity: %w", err)
+	}
+
+	return nil
+}
+
+func (r *DefaultReporter) registerIdentity(identity Identity) error {
 	err := r.cli.Enqueue(segment.Identify{
 		AnonymousId: identity.Installation.ID,
 		Traits:      identity.Installation.TraitsMap(),
@@ -67,18 +94,94 @@ func (r *DefaultReporter) RegisterIdentity(identity Identity) error {
 	return nil
 }
 
-type loggerAdapter struct {
-	log logrus.FieldLogger
+func (r *DefaultReporter) load(ctx context.Context, k8sCli kubernetes.Interface, cfgDir string) (Identity, error) {
+	k8sServerVersion, err := k8sCli.Discovery().ServerVersion()
+	if err != nil {
+		return Identity{}, fmt.Errorf("while getting K8s server version: %w", err)
+	}
+	if k8sServerVersion == nil {
+		return Identity{}, errors.New("server version object cannot be nil")
+	}
+
+	clusterID, err := r.getClusterID(ctx, k8sCli)
+	if err != nil {
+		return Identity{}, fmt.Errorf("while getting cluster ID: %w", err)
+	}
+
+	installationID, err := r.getInstallationID(cfgDir)
+	if err != nil {
+		return Identity{}, fmt.Errorf("while getting installation ID: %w", err)
+	}
+
+	return Identity{
+		Cluster: ClusterIdentity{
+			ID:                clusterID,
+			KubernetesVersion: *k8sServerVersion,
+		},
+		Installation: InstallationIdentity{
+			ID:             installationID,
+			BotKubeVersion: version.Info(),
+		},
+	}, nil
 }
 
-func newLoggerAdapter(log logrus.FieldLogger) *loggerAdapter {
-	return &loggerAdapter{log: log}
+func (r *DefaultReporter) getClusterID(ctx context.Context, k8sCli kubernetes.Interface) (string, error) {
+	kubeSystemNS, err := k8sCli.CoreV1().Namespaces().Get(ctx, kubeSystemNSName, metav1.GetOptions{})
+	if err != nil {
+		return "", fmt.Errorf("while getting %q Namespace: %w", kubeSystemNS, err)
+	}
+	if kubeSystemNS == nil {
+		return "", errors.New("namespace object cannot be nil")
+	}
+
+	return string(kubeSystemNS.GetUID()), nil
 }
 
-func (l *loggerAdapter) Logf(format string, args ...interface{}) {
-	l.log.Infof(format, args) // TODO: Should we use Debug?
+func (r *DefaultReporter) getInstallationID(cfgDir string) (string, error) {
+	analyticsCfgFilePath := filepath.Join(cfgDir, analyticsFileName)
+	if _, err := os.Stat(analyticsCfgFilePath); os.IsNotExist(err) {
+		analyticsCfg := NewConfig()
+		err = r.saveAnalyticsCfg(analyticsCfgFilePath, analyticsCfg)
+		if err != nil {
+			return "", err
+		}
+
+		return analyticsCfg.InstallationID, nil
+	}
+
+	analyticsCfgBytes, err := os.ReadFile(filepath.Clean(analyticsCfgFilePath))
+	if err != nil {
+		return "", fmt.Errorf("while reading analytics config file: %w", err)
+	}
+
+	var analyticsCfg Config
+	err = json.Unmarshal(analyticsCfgBytes, &analyticsCfg)
+	if err != nil {
+		return "", fmt.Errorf("while unmarshalling analytics config file: %w", err)
+	}
+
+	if analyticsCfg.InstallationID == "" {
+		analyticsCfg := NewConfig()
+		err = r.saveAnalyticsCfg(analyticsCfgFilePath, analyticsCfg)
+		if err != nil {
+			return "", err
+		}
+		return analyticsCfg.InstallationID, nil
+	}
+
+	return analyticsCfg.InstallationID, nil
 }
 
-func (l *loggerAdapter) Errorf(format string, args ...interface{}) {
-	l.log.Errorf(format, args)
+func (r *DefaultReporter) saveAnalyticsCfg(path string, cfg Config) error {
+	bytes, err := json.Marshal(cfg)
+	if err != nil {
+		return fmt.Errorf("while marshalling analytics config file: %w", err)
+	}
+
+	err = os.WriteFile(path, bytes, 0600)
+	if err != nil {
+		return fmt.Errorf("while saving analytics config file to %q: %w", path, err)
+	}
+
+	return nil
 }
