+++ internal/analytics/noop_reporter.go
@@ -1,5 +1,10 @@
 package analytics
 
+import (
+	"context"
+	"k8s.io/client-go/kubernetes"
+)
+
 var _ Reporter = &NoopReporter{}
 
 type NoopReporter struct{}
@@ -8,6 +13,6 @@ func NewNoopReporter() *NoopReporter {
 	return &NoopReporter{}
 }
 
-func (n NoopReporter) RegisterIdentity(_ Identity) error {
+func (n NoopReporter) RegisterCurrentIdentity(_ context.Context, _ kubernetes.Interface, _ string) error {
 	return nil
 }
