+++ internal/analytics/reporter.go
@@ -1,5 +1,10 @@
 package analytics
 
+import (
+	"context"
+	"k8s.io/client-go/kubernetes"
+)
+
 type Reporter interface {
-	RegisterIdentity(identity Identity) error
+	RegisterCurrentIdentity(ctx context.Context, k8sCli kubernetes.Interface, cfgDir string) error
 }
